Refactoring and fix mypy [skip ci]
