+++ build_definitions/crypt_blowfish.py
@@ -13,11 +13,10 @@
 #
 
 import os
-import subprocess
 import sys
+import subprocess
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
+from yugabyte_db_thirdparty.builder_interface import BuilderInterface
 from build_definitions import *  # noqa
