+++ build_definitions/glog.py
@@ -12,12 +12,11 @@
 # under the License.
 #
 
-import multiprocessing
 import os
 import sys
+import multiprocessing
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
+from yugabyte_db_thirdparty.builder_interface import BuilderInterface
 from build_definitions import *  # noqa
