+++ build_definitions/hiredis.py
@@ -16,8 +16,7 @@ import os
 import multiprocessing
 import sys
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
+from yugabyte_db_thirdparty.builder_interface import BuilderInterface
 from build_definitions import *  # noqa
