+++ build_definitions/redis_cli.py
@@ -17,8 +17,7 @@ import os
 import sys
 import glob
 
-sys.path.append(os.path.dirname(os.path.dirname(os.path.abspath(__file__))))
-
+from yugabyte_db_thirdparty.builder_interface import BuilderInterface
 from build_definitions import *  # noqa
